feat: preserve directory structure
Used Unfold.ai:

In this updated code, I added a check to see if the `file.name` includes a path using the `std::path::Path::parent()` function. If a path exists, the code creates the required local directory using `fs::create_dir_all()`. Then, the file is downloaded using the `reqwest` crate and saved in the specified directory.

Please note that this code assumes that the file URLs in `file.name` are accessible and can be downloaded.