feat: automatically derive xml URL from details URL
Used ChatGPT-4:

To achieve this in Rust, you'll need to manipulate the URL string to replace the "details" segment with "download". To append the file name to the URL, you can extract the identifier from the original URL and then construct the new URL with this identifier followed by "_files.xml"

First, include the url crate in your Cargo.toml file for URL parsing and manipulation. Use string manipulation functions to replace the relevant part of the URL.

In this code:

replacen("details", "download", 1) replaces the first occurrence of "details" with "download" in the URL. original_url.split('/') splits the URL into segments separated by /.
.last() fetches the last segment, which is the identifier.
format! is used to construct the new URL with the identifier and "_files.xml" appended.