feat: check that URLs are accessible
Used ChatGPT-4:

To write a Rust function that takes a string representing a URL and tests if it is accessible, you will need to use an HTTP client to make a request to the URL and check the response. In Rust, one common way to do this is by using the reqwest crate, which is an asynchronous HTTP client.

The `is_url_accessible()` function does the following:

- It uses `reqwest::Client` to send a GET request to the provided URL.
- If the request is successful and the response status code indicates success (e.g., 200), it returns true.
- If the request is successful but the status code indicates a client or server error (e.g., 404, 500), it returns false.
- If the request fails (e.g., due to a network error), it returns false.
- Note: This function is asynchronous (async). You need to call it within an async context or use an executor like tokio to block on the future.