feat: set download chuck size to 128KB
Used Unfold.ai:

We added a constant `CHUNK_SIZE` which represents the desired chunk size of 128KB. We then created a buffer of size `CHUNK_SIZE` to store the downloaded chunk before writing it to the file. The code uses a loop to handle cases where the received chunk is larger than the desired chunk size, copying and writing chunks of `CHUNK_SIZE` bytes at a time.