fix: correctly resume partial downloads
This took several cycles of asking the AI to correct compiler errors that each proposed solution encountered, until it found one that worked. That said, I have to reorder the code a good bit to ensure `file_size` was available everywhere it was referenced. At this point I have become familiar enough to do that.

Used Unfold.ai:

Based on the provided code snippet, it seems that the file position is being correctly set using the `seek()` method. However, the actual HTTP download may not be starting from the correct offset due to the way the response and chunks are handled.

One possible issue is that the `response.chunk().await?` call might not be resuming the download from the correct offset. This method might not take into account the specified offset and always start from the beginning of the file.

To ensure that the download starts from the correct offset, you can use the `Range` header in the HTTP request to specify the range of bytes you want to download. This can be achieved by modifying the HTTP request headers before sending the request.

To set the `Range` header in `reqwest`, you can use the `header()` method to replace the existing headers with a new `HeaderMap` that contains the `Range` header.